Ignore Missing Ignore/Skip Directives
Currently unused variables are ignored for the purposes of calculating complexity.
However, because graphql-js's `getVariableValues` method returns either a map of coersed
values, OR an array of errors, if directives are present in conjuection with
unused variables then strange errors will be raised.

In particular errors will state that directive variables were not present
when they may well have been.

This change adds fallback measures to ensure that complexity will stil be
calculated if unused variables are present alonside directives.

Fixes #69